Refactor XRay middleware
