+++ examples/security/cmd/multi_auth/http.go
@@ -12,7 +12,8 @@ import (
 	securedservicesvr "goa.design/goa/examples/security/gen/http/secured_service/server"
 	securedservice "goa.design/goa/examples/security/gen/secured_service"
 	goahttp "goa.design/goa/http"
-	"goa.design/goa/http/middleware"
+	httpmiddleware "goa.design/goa/http/middleware"
+	"goa.design/goa/middleware"
 )
 
 // handleHTTPServer starts configures and starts a HTTP server on the given
@@ -63,10 +64,10 @@ func handleHTTPServer(ctx context.Context, u *url.URL, securedServiceEndpoints *
 	var handler http.Handler = mux
 	{
 		if debug {
-			handler = middleware.Debug(mux, os.Stdout)(handler)
+			handler = httpmiddleware.Debug(mux, os.Stdout)(handler)
 		}
-		handler = middleware.Log(adapter)(handler)
-		handler = middleware.RequestID()(handler)
+		handler = httpmiddleware.Log(adapter)(handler)
+		handler = httpmiddleware.RequestID()(handler)
 	}
 
 	// Start HTTP server using default configuration, change the code to
