+++ http/middleware/xray/wrap_doer.go
@@ -5,15 +5,18 @@ import (
 	"net/http"
 
 	goahttp "goa.design/goa/http"
-	"goa.design/goa/http/middleware"
+	"goa.design/goa/middleware"
+	"goa.design/goa/middleware/xray"
 )
 
-// xrayDoer is a goahttp.Doer middleware that will create xray subsegments for traced requests.
+// xrayDoer is a goahttp.Doer middleware that will create xray subsegments for
+// traced requests.
 type xrayDoer struct {
 	wrapped goahttp.Doer
 }
 
-// WrapDoer wraps a goa HTTP Doer and creates xray subsegments for traced requests.
+// WrapDoer wraps a goa HTTP Doer and creates xray subsegments for traced
+// requests.
 func WrapDoer(doer goahttp.Doer) goahttp.Doer {
 	return &xrayDoer{doer}
 }
@@ -22,24 +25,25 @@ func WrapDoer(doer goahttp.Doer) goahttp.Doer {
 func (r *xrayDoer) Do(req *http.Request) (*http.Response, error) {
 	ctx := req.Context()
 
-	seg := ctx.Value(SegKey)
+	seg := ctx.Value(xray.SegKey)
 	if seg == nil {
 		return r.wrapped.Do(req)
 	}
-	s := seg.(*Segment)
+	s := seg.(*xray.Segment)
 	sub := s.NewSubsegment(req.URL.Host)
-	defer sub.Close()
+	hs := &HTTPSegment{Segment: sub}
+	defer hs.Close()
 
 	// update the context with the latest segment
-	ctx = middleware.WithSpan(ctx, sub.TraceID, sub.ID, sub.ParentID)
-	req = req.WithContext(context.WithValue(ctx, SegKey, sub))
+	ctx = middleware.WithSpan(ctx, hs.TraceID, hs.ID, hs.ParentID)
+	req = req.WithContext(context.WithValue(ctx, xray.SegKey, hs.Segment))
 
-	sub.RecordRequest(req, "remote")
+	hs.RecordRequest(req, "remote")
 	resp, err := r.wrapped.Do(req)
 	if err != nil {
-		sub.RecordError(err)
+		hs.RecordError(err)
 	} else {
-		sub.RecordResponse(resp)
+		hs.RecordResponse(resp)
 	}
 	return resp, err
 }
